feat: show permissions for events
